Implement sfUser::load(), add sfUser protected variables
sfUser::load() now correctly loads and stores data via a database query. New protected variables are available for new columns (slug, comments_enabled, and category). This implementation does NOT create a sfUser for the post author, since it may not always be necessary. That data should be lazy-loaded upon request.